+++ src/LocalStorage.js
@@ -1,6 +1,13 @@
 class LocalStorage {
   static get cypressCommands() {
-    return ["clearLocalStorageSnapshot", "saveLocalStorage", "restoreLocalStorage"];
+    return [
+      "clearLocalStorageSnapshot",
+      "saveLocalStorage",
+      "restoreLocalStorage",
+      "setLocalStorage",
+      "getLocalStorage",
+      "removeLocalStorage"
+    ];
   }
 
   constructor(localStorage) {
@@ -24,6 +31,18 @@ class LocalStorage {
     Object.keys(this._snapshot).forEach(key => {
       this._localStorage.setItem(key, this._snapshot[key]);
     });
+  }
+
+  getLocalStorage(key) {
+    return this._localStorage.getItem(key);
+  }
+
+  setLocalStorage(key, value) {
+    return this._localStorage.setItem(key, value);
+  }
+
+  removeLocalStorage(key) {
+    return this._localStorage.removeItem(key);
   }
 }