+++ test-acceptance/app/src/data/user-preferences/selectors.js
@@ -1,0 +1,11 @@
+import { Selector } from "@xbyorange/mercury";
+
+import { userPreferences } from "./origins";
+
+export const cookiesAccepted = new Selector(
+  {
+    source: userPreferences,
+    query: userPreferences.customQueries.cookiesAccepted
+  },
+  cookiesAcceptedResult => cookiesAcceptedResult
+);