+++ test-acceptance/app/src/modules/cookies-value/Controller.js
@@ -1,0 +1,8 @@
+import { connect } from "@xbyorange/react-mercury";
+
+import CookiesValue from "../../components/cookies-value";
+import { cookiesAccepted } from "../../data/user-preferences";
+
+export const Controller = connect(() => ({
+  value: cookiesAccepted.read.getters.value
+}))(CookiesValue);