+++ CHANGELOG.md
@@ -11,6 +11,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 ### BREAKING CHANGES
 
+## [1.1.0] - 2019-10-27
+### Added
+- Add getLocalStorage command
+- Add setLocalStorage command
+- Add removeLocalStorage command
+- Add acceptance tests
+
+### Changed
+- Improve documentation
+
 ## [1.0.0] - 2019-10-26
 ### Added
 - Add saveLocalStorage command